Handle not plain where tree in StorageMerge modifySelect
