+++ tests/queries/0_stateless/01783_merge_engine_join_key_condition.reference
@@ -2,3 +2,4 @@
 1	4
 1	4
 1	4
+1	4
